styled the reviews component; combined RatingsAvgByStar into PrctScaleChar.jsx
