+++ client/src/components/ratings&reviews/PrctScaleChar.jsx
@@ -44,4 +44,37 @@ const PrctScaleChar = ({value}) => {
   )
 };
 
-export default PrctScaleChar;
+const OuterRating = styled.div`
+display: inline-block;
+width: 100px;
+height: 10px;
+border: 1px solid black;
+position: relative;
+background-color: #F8F8F8;
+`;
+
+const InnerRating = styled.div`
+position: absolute;
+top: 0;
+left: 0;
+white-space: nowrap;
+overflow: hidden;
+height: 10px;
+width: ${({percent}) => {
+  return percent;
+}}%;
+background-color: #505050;
+`;
+
+const RatingsAvgByStar = ({value}) => {
+
+
+  return (
+    <OuterRating>
+      <InnerRating percent={value}/>
+    </OuterRating>
+  )
+};
+
+
+export { PrctScaleChar, RatingsAvgByStar };
